move methods only for testing moderation header to InstructorEditStudentFeedbackPage
