+++ src/test/java/teammates/test/pageobjects/FeedbackSubmitPage.java
@@ -1,7 +1,5 @@
 package teammates.test.pageobjects;
 
-import java.io.IOException;
-
 import org.openqa.selenium.By;
 import org.openqa.selenium.JavascriptExecutor;
 import org.openqa.selenium.WebElement;
@@ -199,22 +197,4 @@ public class FeedbackSubmitPage extends AppPage {
         waitForElementToBeClickable(element);
     }
     
-    // ------------- For InstructorEditStudentFeedbackPage -------------
-    
-    public void clickModerationHintButton() {
-        click(By.id("moderationHintButton"));
-    }
-    
-    public boolean isModerationHintVisible() {
-        return isElementVisible("moderationHint");
-    }
-    
-    public void verifyModerationHeaderHtml(String filePathParam) throws IOException {
-        verifyHtmlPart(By.className("navbar"), filePathParam);
-    }
-    
-    public String getModerationHintButtonText() {
-        WebElement moderationHintButton = browser.driver.findElement(By.id("moderationHintButton"));
-        return moderationHintButton.getText();
-    }
 }
